refactor: created IngredientsPickerPage to replace the Popup TODO: revise subscriptions to look at navigable events rather than popup events
