+++ CookNook.csproj
@@ -82,6 +82,9 @@
 	  <Compile Update="Controls\AutocompletePicker.xaml.cs">
 	    <DependentUpon>AutocompletePicker.xaml</DependentUpon>
 	  </Compile>
+	  <Compile Update="IngredientPickerPage.xaml.cs">
+	    <DependentUpon>IngredientPickerPage.xaml</DependentUpon>
+	  </Compile>
 	  <Compile Update="IngredientPickerPopup.xaml.cs">
 	    <DependentUpon>IngredientPickerPopup.xaml</DependentUpon>
 	  </Compile>
@@ -110,6 +113,9 @@
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
 	  <MauiXaml Update="Controls\AutocompletePicker.xaml">
+	    <Generator>MSBuild:Compile</Generator>
+	  </MauiXaml>
+	  <MauiXaml Update="IngredientPickerPage.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
 	  <MauiXaml Update="SearchPage.xaml">
