+++ IngredientPickerPage.xaml
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="CookNook.IngredientPickerPage"
+            xmlns:controls="using:CookNook.Controls">
+    <ContentPage.Content>
+        <controls:AutocompletePicker x:Name="IngredientPicker"/>
+    </ContentPage.Content>
+</ContentPage>
